force integer zooms to keep grid aligned with shot
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@23935 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/apps/elicit/src/zoom.c
@@ -34,7 +34,7 @@ elicit_zoom_add(Evas *evas)
 }
 
 void
-elicit_zoom_zoom_set(Evas_Object *o, double zoom)
+elicit_zoom_zoom_set(Evas_Object *o, int zoom)
 {
   Elicit_Zoom *z;
   int iw, ih;
